Refactor authentication logic to auth.py and update documentation.
Moved authentication code from sow-api/app/main.py to sow-api/app/auth.py.
Updated sow-api/app/main.py and tests to use the new module.
Updated sow-api/README.md, sow-app/README.md, and root README.md with detailed instructions and architecture overview.
Updated .gitignore to exclude Python artifacts.
