+++ .gitignore
@@ -132,3 +132,11 @@ dist
 
 # JetBrains IDEs
 .idea/
+
+# Python
+__pycache__/
+*.py[cod]
+*$py.class
+.pytest_cache/
+.venv
+venv
