+++ sow-api/tests/test_api.py
@@ -1,6 +1,7 @@
 """Unit tests for the sow-api package."""
 from fastapi.testclient import TestClient
-from app.main import app, get_current_user
+from app.main import app
+from app.auth import get_current_user
 from app.models import WaffleReview
 import pytest
